// src/enums/keyboardInput.ts
/**
 * enum mapping key names to KeyboardEvent.code string values
 */
export enum KeyboardInput {
    escape = 'Escape',
    w = 'KeyW',
    a = 'KeyA',
    s = 'KeyS',
    d = 'KeyD',
    shiftLeft = 'ShiftLeft',
    shiftRight = 'shiftRight',
    space = 'Space',
    arrowUp = 'ArrowUp',
    arrowLeft = 'ArrowLeft',
    arrowDown = 'ArrowDown',
    arrowRight = 'ArrowRight'
}